+++ packages/app/package.json
@@ -12,12 +12,12 @@
   },
   "dependencies": {
     "@activitypods/core": "2.0.0-alpha.8",
-    "@activitypods/description": "2.0.0-alpha.8",
     "@rdfjs/data-model": "^1.3.4",
     "@semapps/activitypub": "0.8.8",
     "@semapps/ldp": "0.8.8",
     "@semapps/mime-types": "0.8.8",
     "@semapps/nodeinfo": "0.8.8",
+    "@semapps/ontologies": "0.8.8",
     "form-data-encoder": "1.9.0",
     "formdata-node": "4.4.1",
     "handlebars": "^4.7.7",
