--- packages/description/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ClassDescriptionService: require('./services/class-description'),
-  AccessDescriptionSetService: require('./services/access-description-set')
-};
