--- packages/description/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@activitypods/description",
-  "version": "2.0.0-alpha.8",
-  "description": "ActivityPods package for descriptions",
-  "license": "Apache-2.0",
-  "author": "Virtual Assembly",
-  "homepage": "https://activitypods.org",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/assemblee-virtuelle/activitypods.git",
-    "directory": "packages/core"
-  },
-  "dependencies": {
-    "http-link-header": "^1.1.1",
-    "uuid": "^9.0.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
